requirements.txt
# VeroctaAI Financial Intelligence Platform - Production Requirements

# Web Framework - FastAPI for production
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==23.0.0

# Legacy Flask support (if needed)
flask==3.1.1
werkzeug==3.1.3

# CORS and Authentication
flask-cors==4.0.0
flask-jwt-extended==4.6.0
bcrypt==4.1.2

# Environment Management
python-dotenv==1.0.1

# Data Processing
pandas==2.3.1
numpy>=1.26.0

# AI Integration
openai==1.98.0

# PDF Generation and Visualization
reportlab==4.4.3
matplotlib>=3.5.0
Pillow>=10.0.0

# Date Utilities
python-dateutil>=2.8.2

# Database
psycopg2-binary==2.9.9
sqlalchemy>=2.0.0
flask-sqlalchemy
pydantic>=2.0.0

# HTTP Client
requests==2.31.0

# Payment Processing
stripe==7.8.0

# Email Services
resend==0.8.0
email-validator

# Google Services Integration
google-api-python-client==2.110.0
google-auth-oauthlib==1.2.0
google-auth==2.25.2

# Database Client
supabase==2.3.0
